Updated to use floats

+++ BubblesDivePlanner/ViewModels/Model/Plan/DiveModels/IDiveProfile.cs
@@ -2,16 +2,16 @@ namespace BubblesDivePlanner.ViewModels.Model.Plan.DiveModels
 {
     public interface IDiveProfile
     {
-        double[] MaxSurfacePressures { get; set; }
-        double[] NitrogenTissuePressures { get; set; }
-        double[] HeliumTissuePressures { get; set; }
-        double[] TotalTissuePressures { get; set; }
-        double[] ToleratedAmbientPressures { get; set; }
-        double[] AValues { get; set; }
-        double[] BValues { get; set; }
-        double[] CompartmentLoads { get; set; }
-        double OxygenAtPressure { get; set; }
-        double NitrogenAtPressure { get; set; }
-        double HeliumAtPressure { get; set; }
+        float[] MaxSurfacePressures { get; set; }
+        float[] NitrogenTissuePressures { get; set; }
+        float[] HeliumTissuePressures { get; set; }
+        float[] TotalTissuePressures { get; set; }
+        float[] ToleratedAmbientPressures { get; set; }
+        float[] AValues { get; set; }
+        float[] BValues { get; set; }
+        float[] CompartmentLoads { get; set; }
+        float OxygenAtPressure { get; set; }
+        float NitrogenAtPressure { get; set; }
+        float HeliumAtPressure { get; set; }
     }
 }
